// web-application/src/types-openapi/api/models/ListAdministratorRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListAdministratorRequest
 */
export interface ListAdministratorRequest {
    /**
     * 
     * @type {string}
     * @memberof ListAdministratorRequest
     */
    keyword?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListAdministratorRequest
     */
    role?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListAdministratorRequest
     */
    status?: string | null;
}

/**
 * Check if a given object implements the ListAdministratorRequest interface.
 */
export function instanceOfListAdministratorRequest(value: object): value is ListAdministratorRequest {
    return true;
}

export function ListAdministratorRequestFromJSON(json: any): ListAdministratorRequest {
    return ListAdministratorRequestFromJSONTyped(json, false);
}

export function ListAdministratorRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListAdministratorRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'keyword': json['keyword'] == null ? undefined : json['keyword'],
        'role': json['role'] == null ? undefined : json['role'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function ListAdministratorRequestToJSON(json: any): ListAdministratorRequest {
    return ListAdministratorRequestToJSONTyped(json, false);
}

export function ListAdministratorRequestToJSONTyped(value?: ListAdministratorRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'keyword': value['keyword'],
        'role': value['role'],
        'status': value['status'],
    };
}