// web-application/src/types-openapi/api/models/ListUserRoleItemResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListUserRoleItemResponse
 */
export interface ListUserRoleItemResponse {
    /**
     * 
     * @type {string}
     * @memberof ListUserRoleItemResponse
     */
    roleId: string;
    /**
     * 
     * @type {string}
     * @memberof ListUserRoleItemResponse
     */
    roleName: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof ListUserRoleItemResponse
     */
    assigned: boolean;
}

/**
 * Check if a given object implements the ListUserRoleItemResponse interface.
 */
export function instanceOfListUserRoleItemResponse(value: object): value is ListUserRoleItemResponse {
    if (!('roleId' in value) || value['roleId'] === undefined) return false;
    if (!('roleName' in value) || value['roleName'] === undefined) return false;
    if (!('assigned' in value) || value['assigned'] === undefined) return false;
    return true;
}

export function ListUserRoleItemResponseFromJSON(json: any): ListUserRoleItemResponse {
    return ListUserRoleItemResponseFromJSONTyped(json, false);
}

export function ListUserRoleItemResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListUserRoleItemResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'roleId': json['roleId'],
        'roleName': json['roleName'],
        'assigned': json['assigned'],
    };
}

export function ListUserRoleItemResponseToJSON(json: any): ListUserRoleItemResponse {
    return ListUserRoleItemResponseToJSONTyped(json, false);
}

export function ListUserRoleItemResponseToJSONTyped(value?: ListUserRoleItemResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'roleId': value['roleId'],
        'roleName': value['roleName'],
        'assigned': value['assigned'],
    };
}